--- engines/query-sparql/test/service-description-source.ttl
@@ -1,3 +1,0 @@
-@prefix ex: <http://example.org/> .
-
-ex:subject ex:predicate ex:object .
